fix: 기능별 분할
